Mesh adaptivity

FEMSolids.refine — Function

refine(grid::Grid, cells_to_split)

Refine the given grid by applying the Rivara algorithm where cells_to_split is the list of cells to define (should contain cell numbers). This implementation is restricted to linear 2D Triangles.

Warning

nodesets, facesets and cellsets are currently lost after refinement and must be reconstructed. They can be added to a grid in the following manner:

addfaceset!(mesh, "left", x -> x[1] ≈ 0.0)
addfaceset!(mesh, "bottom", x -> x[2] ≈ 0.0)
addfaceset!(mesh, "Γₗ", x -> (x[1] <= h && x[2] ≈ h))

FEMSolids.transfer_solution — Function

transfer_solution(dh_from::DofHandler, dh_to::DofHandler, a_from::AbstractVector)

Interpolate the solution a_from (associated with dh_from) to the degrees of freedom of the DofHandler dh_to. Return the interpolated values as a vector ordered according to the dofs in dh_to.