Old code from previous course revisions

Equivalent functionality is now documented provided in Ferrite v1. Included here for legacy reasons.

FEMSolids.LinearElasticity — Type

LinearElasticity{dim}(;G, K) where dim

Construct a LinearElasticity material based on the shear modulus G and the bulk modulus K.

FEMSolids.Neumann — Type

Neumann(f::Function, facetset::AbstractSet{FacetIndex}, nfacets::Int)

Represent a Neumann boundary condition.

Arguments:

f: Prescribed traction as a function f(x,t), with spatial coordinate x and time t

facetset: The facets on which the boundary condition acts

nfacets: Number of facets per cell

Example:

grid = generate_grid(Quadrilateral, (5,5))

prescribed_traction(x,t) = Vec(zero(t), x[1]*t)
facetset = getfacetset(grid, "top")
nfacets = nfacets(getcells(grid, 1))

neumann_bc = Neumann(prescribed_traction, facetset, nfacets)

Note

Neumann needs to be updated once for each time step with update! and once for each cell with update_cell!.

FEMSolids.Primal — Type

Primal(material, thickness::Float64)

Represent the weak form of the moment equilibrium in primal format:

\[\int_\Omega \boldsymbol{\sigma} : \boldsymbol{\varepsilon}(\delta \boldsymbol{u}) \, \mathrm{d}\Omega = \int_\Omega \boldsymbol{b} \cdot \delta\boldsymbol{u} \, \mathrm{d}\Omega + \int_{\partial\Omega_\mathrm{N}} \boldsymbol{t}_\mathrm{p} \cdot \delta\boldsymbol{u} \, \mathrm{d}\Gamma,\]

where $\boldsymbol{\sigma}$ is the stress tensor, $ \boldsymbol{\varepsilon} $ is the strain tensor, $ \boldsymbol{b} $ is the vector of body forces, $ \boldsymbol{t}_\mathrm{p} $ is the prescribed boundary traction and $ \delta \boldsymbol{u} $ is the test function.

Boundary conditions on the Dirichlet boundary $\Gamma_\text{D}$ and the Neumann boundary $\Gamma_\text{N}$ are respectively given as:

\[\begin{aligned} \boldsymbol{u} &= \boldsymbol{u}_\text{p} \quad \text{on} \quad \Gamma_\text{D} \\ \boldsymbol{\sigma} \cdot \boldsymbol{n} &= \boldsymbol{t}_\text{p} \quad \text{on} \quad \Gamma_\text{N} \\ \end{aligned}\]

where $\boldsymbol{u}$ is the displacement field.

Arguments:

materal: Material model for the stress-strain relation

thickness: Cross-sectional area / out-of-plane thickness for 1D / 2D problems

FEMSolids.element_routine! — Function

element_routine!(weak_form::Primal, ke, fe, fe_external, cv, xe, ue [, fv, neumann_bc::Neumann])

Compute the element stiffness matrix ke, the internal force vector fe and the external force vector fe_external:

\[\begin{aligned} \left( \texttt{ke} \right)_{ij} &= a\left( \boldsymbol{N}^{(\rm u)}_i, \, \boldsymbol{N}^{(\rm u)}_j \right) = \int_\Omega \boldsymbol{\nabla}^\text{sym} \boldsymbol{N}^{(\rm u)}_i : \boldsymbol{\mathsf{E}} : \boldsymbol{\nabla}^\text{sym} \boldsymbol{N}^{(\rm u)}_j \rm d \Omega \\ \left( \texttt{fe} \right)_{i} &= a\left( \boldsymbol{u}, \, \boldsymbol{N}^{(\rm u)}_j \right) = \int_\Omega \boldsymbol{\nabla}^\text{sym} \boldsymbol{u} : \boldsymbol{\mathsf{E}} : \boldsymbol{\nabla}^\text{sym} \boldsymbol{N}^{(\rm u)}_j \rm d \Omega \\ \left( \texttt{fe\_external} \right)_{i} &= l\left( \boldsymbol{N}^{(\rm u)}_i \right) = \int_{\Gamma^{(\rm u)}_\text{N} } \boldsymbol{N}^{(\rm u)}_i \cdot \boldsymbol{t}_\text{p} \, \rm d \Gamma, \end{aligned}\]

where $\boldsymbol{N}^{(\rm u)}_i$ is the shape function for the i-th degree of freedom, $ \mathsf{E} $ is the 4-th order elastic stiffness tensor, $\boldsymbol{u}$ is the displacement and $\boldsymbol{t}_\mathrm{p}$ is the prescribed traction.

Arguments:

weak_form: weak form representation

ke: element stiffness matrix

fe: element force vector

cv: CellValues

xe: element coordinate vector

ue: element displacement vector

Optional arguments for integrating the load vector:

fv: FaceVectorValues

neumann_bc: Neumann boundary condition

FEMSolids.update_cell! — Function

update_cell!(bc::Neumann, cellid::Int)

Update the Neumann boundary condition bc with the current cellid.

Ferrite.update! — Method

update!(bc::Neumann, time::Float64=0.0)

Update the Neumann boundary condition bc for the current time.