FEMSolids

Documentation for FEMSolids.

This package provides routines that are needed for solving the assignments in FEM Solids.

FEMSolids.LinearElasticity

FEMSolids.Primal

FEMSolids.element_routine!

FEMSolids.linear_to_quadratic

FEMSolids.refine

FEMSolids.LinearElasticity — Method

LinearElasticity{dim}(;G, K) where dim

Construct a LinearElasticity material based on the shear modulus G and the bulk modulus K.

FEMSolids.Primal — Type

Primal

This struct is used for dispatching which element_routine! is used when you add your own element routines. Primal refers to the primal format of the moment equilibrium:

\[- \boldsymbol{\sigma} \cdot \boldsymbol{\nabla} = \boldsymbol{0}\]

with boundary conditions on the boundary $\Gamma = \Gamma_\text{D} \cup \Gamma_\text{N}$

\[\begin{aligned} \boldsymbol{u} &= \boldsymbol{u}_\text{p} \quad \text{on} \quad \Gamma_\text{D} \\ \boldsymbol{t} &= \boldsymbol{t}_\text{p} \quad \text{on} \quad \Gamma_\text{N} \\ \end{aligned}\]

where $\boldsymbol{u}$ is the primary unknown field.

FEMSolids.element_routine! — Function

element_routine!(::Primal, ke, fe, cv, xe, material, thickness[, fv, grid, cellid, tₚ, faceset_name])

Compute the element stiffness matrix ke and the element external load vector fe:

\[\begin{aligned} \left( ke \right)_{ij} &= a\left( \boldsymbol{N}^{(\rm u)}_i, \, \boldsymbol{N}^{(\rm u)}_j \right) = \int_\Omega \boldsymbol{\nabla}^\text{sym} \boldsymbol{N}^{(\rm u)}_i : \boldsymbol{\mathsf{E}} : \boldsymbol{\nabla}^\text{sym} \boldsymbol{N}^{(\rm u)}_j \rm d \Omega \\ \left( fe \right)_{i} &= l\left( \boldsymbol{N}^{(\rm u)}_i \right) = \int_{\Gamma^{(\rm u)}_\text{N} } \boldsymbol{N}^{(\rm u)}_i \cdot \boldsymbol{t}_\text{p} \, \rm d \Gamma \end{aligned}\]

Arguments:

ke: element stiffness matrix

fe: element force vector

cv: CellVectorValues

xe: element coordinate vector

material: material parameters, for instant only LinearElasticity is allowed

thickness: out-of-plane thickness

Optional arguments for integrating the load vector:

fv: FaceVectorValues

grid: Grid

cellid: global cell index of the current cell

tₚ: traction vector

faceset_name: name of the faceset over which the load vector should be integrated

FEMSolids.linear_to_quadratic — Method

linear_to_quadratic(dh_lin::DofHandler{2,Triangle}, dh_quad::DofHandler{2,Triangle}, a_lin)

Interpolate the linear solution a_lin (associated with dh_lin) to the degrees of freedom of the quadratic DofHandler dh_quad. Return the interpolated values as a vector ordered according to the dofs in dh_quad. Restricted to vector valued fields.

FEMSolids.refine — Method

refine(grid::Grid, cells_to_split)

Refine the given grid by applying the Rivara algorithm where cells_to_split is the list of cells to define (should contain cell numbers). The grid must include a Topology for the refinement to work. Construct the first grid as follows, all grids that are returned by refine include the topology information:

grid = generate_grid(Triangle, (nel_x, nel_y), lower_left, upper_right; ; build_topology=true)

This implementation is restricted to linear 2D Triangles.

Warning

nodesets, facesets and cellsets are currently lost after refinement and must be reconstructed. They can be added to a grid in the following manner:

addfaceset!(mesh, "left", x -> x[1] ≈ 0.0)
addfaceset!(mesh, "bottom", x -> x[2] ≈ 0.0)
addfaceset!(mesh, "Γₗ", x -> (x[1] <= h && x[2] ≈ h))